refactor(canvas): extract chat session domain
