+++ apps/canvas-workspace/AGENTS.md
@@ -88,7 +88,7 @@ deploys the external-agent `pulse-canvas` CLI + bundled skills. Do not mix them.
 | Diagnose Canvas Agent latency or verify an optimization | `harness/skills/diagnose-agent-latency/SKILL.md`; trace model and exporter details: `harness/knowledge/langfuse-observability.md` |
 | Canvas Agent full-chain Langfuse observability | `harness/knowledge/langfuse-observability.md` |
 | xterm sizing for agent/terminal surfaces | Read `harness/knowledge/terminal-surfaces.md` before changing xterm sizing/fit behavior for agent nodes, terminal nodes, or the workspace terminal dock. Key contracts: `AgentNodeBody/utils/terminal.ts`, `TerminalNodeBody/index.tsx`, `WorkspaceTerminalDock/index.tsx`. Tests: `AgentNodeBody/utils/terminalFit.test.ts` |
-| AI Chat loading states | Read `harness/knowledge/chat-sessions.md` before changing session loading flags (`sessionsLoading` vs `sessionLoading` vs stream `loading`) or the submit veto. Key contracts: `src/renderer/src/components/chat/hooks/useChatSessions.ts`, `hooks/useMentions.ts`, `ChatSessionsRail.tsx`, `ChatHeader.tsx`, `ChatThreadSkeleton.tsx`. Tests: `hooks/useChatSessions.test.tsx`, `hooks/useMentions.submit-veto.test.tsx`, `__tests__/ChatSessionLoading.test.tsx` |
+| AI Chat loading states | Read `harness/knowledge/chat-sessions.md` before changing session loading flags (`sessionsLoading` vs `sessionLoading` vs stream `loading`) or the submit veto. Key contracts: `src/renderer/src/agent-chat/sessions/useChatSessions.ts`, `components/chat/hooks/useMentions.ts`, `components/chat/ChatSessionsRail.tsx`, `components/chat/ChatHeader.tsx`, `components/chat/ChatThreadSkeleton.tsx`. Tests: `agent-chat/sessions/useChatSessions.test.tsx`, `components/chat/hooks/useMentions.submit-veto.test.tsx`, `components/chat/__tests__/ChatSessionLoading.test.tsx` |
 | Full-page chat ↔ dock Tabs | Read `harness/knowledge/chat-sessions.md` before changing the full-page chat topbar, its dock-content-tabs toggle, or the dock inset rule. Key contracts: `chat/ChatPageBody.tsx`, `RightDock/dock-content-tabs.ts`. Tests: `RightDock/__tests__/dock-content-tabs.test.ts`, `RightDock/index.test.tsx` |
 | Dock width policy | Read `harness/knowledge/chat-sessions.md` before changing right-dock width behavior. Key contract: `RightDock/dock-width.ts`. Tests: `RightDock/__tests__/dock-width.test.ts`, `RightDock/index.test.tsx` |
 | Coding agents in the Coding Agent node (roster, brand marks, launch flags, per-node session binding) | Read `harness/knowledge/coding-agent-registry.md` before adding/removing an agent, changing per-agent launch flags, or changing how a node binds to one CLI conversation. Key contracts: `src/renderer/src/config/agentRegistry.ts`, `AgentNodeBody/utils/piSession.ts`, `src/main/agent/codex-sessions.ts`. Bound tests (`coding-agent-roster` rule, `harness/validate/validation.yaml`): `AgentNodeBody/__tests__/AgentPicker.test.tsx`, `AgentNodeBody/__tests__/AgentIcon.test.tsx`, `AgentNodeBody/__tests__/piSessionBinding.test.tsx`, `AgentNodeBody/utils/piSession.test.ts`, `utils/__tests__/codingAgentCommand.test.ts`. |
@@ -156,7 +156,7 @@ deploys the external-agent `pulse-canvas` CLI + bundled skills. Do not mix them.
   Guard: `segment-execution.test.ts`, `chat-stop.test.ts`.
   Detail: `harness/knowledge/agent-roles.md` (stopped-vs-failed turn rule).
 - The full-screen chat rail is one stable cross-scope projection: never swap per-scope list caches into it or fetch a list while `loadSession` is promoting an archive; commit current/other lists together after promotion.
-  Guard: `useChatSessions.test.tsx`, `ChatSessionsRail.test.tsx`.
+  Guard: `src/renderer/src/agent-chat/sessions/useChatSessions.test.tsx`, `ChatSessionsRail.test.tsx`.
   Detail: `harness/knowledge/chat-sessions.md`.
 - Chat-target registration is synchronously observed at the app root: props feeding a mounted `ChatPanel` target or its handlers must use stable empty-collection fallbacks, never an inline `[]`, or the target unregisters/re-registers every root render and hits React's max update depth.
   Guard: `Workbench/__tests__/ChatDockLifecycle.test.tsx` (knowledge chat follows the same rule but isn't covered by this guard).
